Allow Fog Range to be decreased further

+++ D3D11Engine/GothicGraphicsState.h
@@ -719,7 +719,7 @@ struct GothicRendererSettings {
     bool DrawParticleEffects;
     bool DrawSky;
     bool DrawFog;
-    int FogRange;
+    float FogRange;
     int WindQuality;
     bool HeroAffectsObjects;
     bool DrawG1ForestPortals;
